MSVC warnings fixed

+++ modules/rgbd/src/tsdf.cpp
@@ -1387,9 +1387,9 @@ void TSDFVolumeGPU::fetchPointsNormals(OutputArray points, OutputArray normals) 
         size_t lplanes = min(memSize/elemSize, wgsLimit)/lcols/lrows;
         lplanes = roundDownPow2(lplanes);
         size_t localSize[3] = {lcols, lrows, lplanes};
-        Vec3i ngroups((int)divUp(globalSize[0], (size_t)localSize[0]),
-                      (int)divUp(globalSize[1], (size_t)localSize[1]),
-                      (int)divUp(globalSize[2], (size_t)localSize[2]));
+        Vec3i ngroups((int)divUp(globalSize[0], (unsigned int)localSize[0]),
+                      (int)divUp(globalSize[1], (unsigned int)localSize[1]),
+                      (int)divUp(globalSize[2], (unsigned int)localSize[2]));
 
         const size_t counterSize = sizeof(int);
         size_t lsz = localSize[0]*localSize[1]*localSize[2]*counterSize;
@@ -1415,7 +1415,7 @@ void TSDFVolumeGPU::fetchPointsNormals(OutputArray points, OutputArray normals) 
             throw std::runtime_error("Failed to run kernel");
 
         Mat groupedSumCpu = groupedSum.getMat(ACCESS_READ);
-        int gpuSum = cv::sum(groupedSumCpu)[0];
+        int gpuSum = (int)cv::sum(groupedSumCpu)[0];
         // should be no CPU copies when new kernel is executing
         groupedSumCpu.release();
